Igo gauzak

+++ src/main/java/org/osb/web/domain/ikasgaia/repository/IkasgaiaRepository.java
@@ -2,10 +2,11 @@ package org.osb.web.domain.ikasgaia.repository;
 
 import java.util.Optional;
 
-import org.osb.web.domain.ikasgaia.model.Ikasgaia;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IkasgaiaRepository extends CrudRepository<Ikasgaia, Integer> {
+import org.osb.web.domain.ikasgaia.model.Ikasgaia;
+
+public interface IkasgaiaRepository extends CrudRepository<Ikasgaia, Long> {
     
     Optional<Ikasgaia> findByIzena(String izena);
